requirements.txt
selenium~=4.15.2
webdriver-manager
python-telegram-bot==20.6
asyncio
python-dotenv